auth/requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.2
PyYAML==3.12
requests==2.17.3
requests-oauth==0.4.1
requests-oauthlib==0.8.0
SQLAlchemy==1.1.10
urllib3==1.21.1
Werkzeug==0.12.2
